requirements.txt
streamlit
sentence-transformers
faiss-cpu
pypdf
groq
numpy
